Update GCP authentication documentation (#12)
## What documentation does this update and for what reason?
This PR updates the documentation for authenticating with GCP via the `gcloud` CLI  as opposed to creating a service account, etc. This is done so that users do not have to manage sensitive security keys that could potentially get leaked onto a repository commit. It ALSO further standardizes the formatting of the markdown files and gets rid of image files that are no longer being used.